Added Info tool tip for quality indicators in submit form
